将各结构的方法归位并重新划分了包，增加了命令行模式并支持将收到的礼物信息、弹幕信息通过POST请求回调到指定的地址以供第三方程序处理
