+++ handler/getinfo.go
@@ -1,4 +1,4 @@
-package models
+package handler
 
 import (
 	"bytes"
@@ -9,10 +9,12 @@ import (
 	"io/ioutil"
 	"math"
 	"net/http"
+
+	"biliDanMu/models"
 )
 
 func GetRealRoomID(short int) (realID uint32, err error) {
-	url := fmt.Sprintf("%s?id=%d", RealID, short)
+	url := fmt.Sprintf("%s?id=%d", models.RealID, short)
 	resp, err := http.Get(url)
 	if err != nil {
 		fmt.Println("http.Get token err: ", err)
@@ -25,14 +27,14 @@ func GetRealRoomID(short int) (realID uint32, err error) {
 		fmt.Println("ioutil.ReadAll(resp.Body) err: ", err)
 		return 0, err
 	}
-	realID = json.Get(rawdata, "data", "room_id").ToUint32()
+	realID = models.Json.Get(rawdata, "data", "room_id").ToUint32()
 
 	return realID, nil
 }
 
 // GetToken return the necessary token for connecting to the server
 func GetToken(roomid uint32) (key string) {
-	url := fmt.Sprintf("%s?room_id=%d&platform=pc&player=web", keyUrl, roomid)
+	url := fmt.Sprintf("%s?room_id=%d&platform=pc&player=web", models.KeyUrl, roomid)
 
 	resp, err := http.Get(url)
 	if err != nil {
@@ -47,13 +49,13 @@ func GetToken(roomid uint32) (key string) {
 		fmt.Println("ioutil.ReadAll(resp.Body) err: ", err)
 		return
 	}
-	key = json.Get(rawdata, "data").Get("token").ToString()
+	key = models.Json.Get(rawdata, "data").Get("token").ToString()
 	return
 }
 
 func GetRoomInfo(roomid uint32) (roomInfo RoomInfo) {
 	// roomInfo = &models.RoomInfo{}
-	url := fmt.Sprintf("%s?room_id=%d", roomInfoUrl, roomid)
+	url := fmt.Sprintf("%s?room_id=%d", models.RoomInfoUrl, roomid)
 
 	resp, err := http.Get(url)
 	if err != nil {
@@ -70,11 +72,11 @@ func GetRoomInfo(roomid uint32) (roomInfo RoomInfo) {
 	}
 
 	roomInfo.RoomId = roomid
-	roomInfo.UpUid = json.Get(rawdata, "data").Get("room_info").Get("uid").ToUint32()
-	roomInfo.Title = json.Get(rawdata, "data").Get("room_info").Get("title").ToString()
-	roomInfo.Tags = json.Get(rawdata, "data").Get("room_info").Get("tags").ToString()
-	roomInfo.LiveStatus = json.Get(rawdata, "data").Get("room_info").Get("live_status").ToBool()
-	roomInfo.LockStatus = json.Get(rawdata, "data").Get("room_info").Get("lock_status").ToBool()
+	roomInfo.UpUid = models.Json.Get(rawdata, "data").Get("room_info").Get("uid").ToUint32()
+	roomInfo.Title = models.Json.Get(rawdata, "data").Get("room_info").Get("title").ToString()
+	roomInfo.Tags = models.Json.Get(rawdata, "data").Get("room_info").Get("tags").ToString()
+	roomInfo.LiveStatus = models.Json.Get(rawdata, "data").Get("room_info").Get("live_status").ToBool()
+	roomInfo.LockStatus = models.Json.Get(rawdata, "data").Get("room_info").Get("lock_status").ToBool()
 
 	return
 }
@@ -92,27 +94,6 @@ func ZlibInflate(compress []byte) ([]byte, error) {
 		return out.Bytes(), nil
 	}
 	return nil, err
-}
-
-func (d *DanMuMsg) GetDanmuMsg(source []byte) {
-	d.UID = json.Get(source, "info", 2, 0).ToUint32()
-	d.Uname = json.Get(source, "info", 2, 1).ToString()
-	d.Ulevel = json.Get(source, "info", 4, 0).ToUint32()
-	d.Text = json.Get(source, "info", 1).ToString()
-	d.MedalName = json.Get(source, "info", 3, 1).ToString()
-	if d.MedalName == "" {
-		d.MedalName = "无勋章"
-	}
-	d.MedalLevel = json.Get(source, "info", 3, 0).ToUint32()
-	return
-}
-
-func (g *Gift) GetGiftMsg(source []byte) {
-	g.UUname = json.Get(source, "data", "uname").ToString()
-	g.Action = json.Get(source, "data", "action").ToString()
-	nums := json.Get(source, "data", "num").ToUint32()
-	g.Price = json.Get(source, "data", "price").ToUint32() * nums
-	g.GiftName = json.Get(source, "data", "giftName").ToString()
 }
 
 // 返回字节数组表示数的十进制形式
